svg import : store font size as a double + some minor refactoring
git-svn-id: svn://scribus.net/trunk/Scribus@17436 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/plugins/import/svg/svgplugin.h
@@ -176,7 +176,7 @@ public:
 	QString FontStyle;
 	QString FontWeight;
 	QString FontStretch;
-	int FontSize;
+	double  FontSize;
 	QString FillCol;
 	QString fillRule;
 	QString GFillCol1;
